+++ src/subatomic_species.jl
@@ -77,18 +77,18 @@ function subatomic_particle(name::String)
       particle.mass,
       particle.spin,
       particle.moment,
-      0, Kind.PHOTON)
+      Int32(0), Kind.PHOTON)
   elseif name in leptons
     return Species(name, particle.charge,
       particle.mass,
       particle.spin,
       particle.moment,
-      0, Kind.LEPTON)
+      Int32(0), Kind.LEPTON)
   else
     return Species(name, particle.charge,
       particle.mass,
       particle.spin,
       particle.moment,
-      0, Kind.HADRON)
+      Int32(0), Kind.HADRON)
   end
 end
